requirements.txt
pycryptodome
# build
pyinstaller